cloud_libraries/spanner/requirements.txt
flask
google-cloud-spanner
gunicorn
